ENH docs partially updated, happy mooncake festival
