+++ sklearn/manifold/tests/test_spectral_embedding.py
@@ -1,63 +1,27 @@
 from nose.tools import assert_true
 from nose.tools import assert_equal
 
-from scipy import sparse
 from scipy.sparse import csr_matrix
 import numpy as np
 from numpy.testing import assert_almost_equal, assert_array_almost_equal
 
-from nose.tools import assert_raises
-
-from sklearn.manifold.spectral_embedding import SpectralEmbedding
+from sklearn.manifold.spectra_embedding import SpectralEmbedding
 from sklearn.metrics.pairwise import rbf_kernel
 from sklearn.pipeline import Pipeline
 from sklearn.metrics import normalized_mutual_info_score
 from sklearn.cluster import KMeans, SpectralClustering
-from sklearn.datasets.samples_generator import make_blobs
+
+S = np.array([[1, 5, 2, 1, 0, 0, 0],
+              [5, 1, 3, 1, 0, 0, 0],
+              [2, 3, 1, 1, 0, 0, 0],
+              [1, 1, 1, 1, 2, 1, 1],
+              [0, 0, 0, 2, 2, 3, 2],
+              [0, 0, 0, 1, 3, 1, 4],
+              [0, 0, 0, 1, 2, 4, 1],
+              ])
 
 
-# non centered, sparse centers to check the
-centers = np.array([
-    [0.0, 5.0, 0.0, 0.0, 0.0],
-    [0.0, 0.0, 4.0, 0.0, 0.0],
-    [1.0, 0.0, 0.0, 5.0, 1.0],
-])
-n_samples = 1000
-n_clusters, n_features = centers.shape
-S, true_labels = make_blobs(n_samples=n_samples, centers=centers,
-                            cluster_std=1., random_state=42)
-
-
-def test_spectral_embedding_two_components(seed=36):
-    """Test spectral embedding with two components"""
-    random_state = np.random.RandomState(seed)
-    n_sample = 10
-    affinity = np.zeros(shape=[n_sample * 2,
-                               n_sample * 2])
-    # first component
-    affinity[0:n_sample,
-             0:n_sample] = np.abs(random_state.randn(n_sample, n_sample)) + 2
-    # second component
-    affinity[n_sample::,
-             n_sample::] = np.abs(random_state.randn(n_sample, n_sample)) + 2
-    affinity = 0.5 * (affinity + affinity.T)
-    # connection
-    affinity[0, n_sample + 1] = 1
-    affinity[n_sample + 1, 0] = 1
-    affinity.flat[::2 * n_sample + 1] = 0
-
-    true_label = np.zeros(shape=2 * n_sample)
-    true_label[0:n_sample] = 1
-
-    se_precomp = SpectralEmbedding(n_components=1, affinity="precomputed",
-                                   random_state=np.random.RandomState(seed))
-    embedded_corrdinate = np.squeeze(se_precomp.fit_transform(affinity))
-    # thresholding on the first components using 0.
-    label_ = np.array(embedded_corrdinate < 0, dtype="float")
-    assert_equal(normalized_mutual_info_score(true_label, label_), 1.0)
-
-
-def test_spectral_embedding_precomputed_affinity(seed=36):
+def test_spectra_embedding_precomputed_graph(seed=36):
     """Test spectral embedding with precomputed kernel"""
     gamma = 1.0
     se_precomp = SpectralEmbedding(n_components=3, affinity="precomputed",
@@ -65,15 +29,23 @@ def test_spectral_embedding_precomputed_affinity(seed=36):
     se_rbf = SpectralEmbedding(n_components=3, affinity="rbf",
                                gamma=gamma,
                                random_state=np.random.RandomState(seed))
+    se_knn = SpectralEmbedding(n_components=3, affinity="nearest_neighbors",
+                               gamma=gamma,
+                               random_state=np.random.RandomState(seed))
     embed_precomp = se_precomp.fit_transform(rbf_kernel(S, gamma=gamma))
     embed_rbf = se_rbf.fit_transform(S)
+    embed_knn = se_knn.fit_transform(S)
     assert_array_almost_equal(
         se_precomp.affinity_matrix_, se_rbf.affinity_matrix_)
     assert_array_almost_equal(np.abs(embed_precomp), np.abs(embed_rbf), 0)
 
 
-def test_spectral_embedding_callable_affinity(seed=36):
-    """Test spectral embedding with callable affinity"""
+def test_spectra_embedding_knn_graph(seed=36):
+    """Test spectral embedding with knn graph"""
+
+
+def test_spectra_embedding_callable_graph(seed=36):
+    """Test spectral embedding with knn graph"""
     gamma = 0.9
     kern = rbf_kernel(S, gamma=gamma)
     se_callable = SpectralEmbedding(n_components=3,
@@ -86,76 +58,24 @@ def test_spectral_embedding_callable_affinity(seed=36):
                                random_state=np.random.RandomState(seed))
     embed_rbf = se_rbf.fit_transform(S)
     embed_callable = se_callable.fit_transform(S)
-    embed_rbf = se_rbf.fit_transform(S)
-    embed_callable = se_callable.fit_transform(S)
-    assert_array_almost_equal(
-        se_callable.affinity_matrix_, se_rbf.affinity_matrix_)
-    assert_array_almost_equal(np.abs(embed_rbf), np.abs(embed_callable), 2)
 
 
-def test_spectral_embedding_amg_solver(seed=36):
-    """Test spectral embedding with amg solver"""
-    gamma = 0.9
-    se_amg = SpectralEmbedding(n_components=3, affinity="rbf",
-                               gamma=gamma, eigen_solver="amg",
-                               random_state=np.random.RandomState(seed))
-    se_arpack = SpectralEmbedding(n_components=3, affinity="rbf",
-                                  gamma=gamma, eigen_solver="arpack",
-                                  random_state=np.random.RandomState(seed))
-    embed_amg = se_amg.fit_transform(S)
-    embed_arpack = se_arpack.fit_transform(S)
-    assert_array_almost_equal(
-        se_amg.affinity_matrix_, se_arpack.affinity_matrix_)
-    assert_array_almost_equal(np.abs(embed_amg), np.abs(embed_arpack), 2)
-
-
-def test_pipline_spectral_clustering(seed=36):
+def test_pipline_spectra_clustering():
     """Test using pipline to do spectral clustering"""
-    random_state = np.random.RandomState(seed)
-    se_rbf = SpectralEmbedding(n_components=n_clusters,
-                               affinity="rbf",
-                               random_state=random_state)
-    se_knn = SpectralEmbedding(n_components=n_clusters,
-                               affinity="nearest_neighbors",
-                               n_neighbors=5,
-                               random_state=random_state)
-    for se in [se_rbf, se_knn]:
-        km = KMeans(n_clusters=n_clusters, random_state=random_state)
-        km.fit(se.fit_transform(S))
+    spectral_clustering = Pipeline([
+        ('se', SpectralEmbedding()),
+        ('km', KMeans()),
+    ])
+
+    for n_cluster in range(1, 5):
+        n_cluster = 3
+        spectral_clustering.set_params(km__n_clusters=n_cluster)
+        spectral_clustering.set_params(se__n_components=n_cluster)
+        spectral_clustering.set_params(se__gamma=1.0)
+        spectral_clustering.fit(S)
+        SC = SpectralClustering(n_clusters=n_cluster)
+        SC.fit(S)
         assert_array_almost_equal(
             normalized_mutual_info_score(
-                km.labels_,
-                true_labels), 1.0, 2)
-
-
-def test_spectral_embedding_unknown_eigensolver(seed=36):
-    """Test that SpectralClustering fails with an unknown eigensolver"""
-    centers = np.array([
-        [0., 0., 0.],
-        [10., 10., 10.],
-        [20., 20., 20.],
-    ])
-    X, true_labels = make_blobs(n_samples=100, centers=centers,
-                                cluster_std=1., random_state=42)
-    D = rbf_kernel(X)  # Distance matrix
-
-    se_precomp = SpectralEmbedding(n_components=1, affinity="precomputed",
-                                   random_state=np.random.RandomState(seed),
-                                   eigen_solver="<unknown>")
-    assert_raises(ValueError, se_precomp.fit, S)
-
-
-def test_spectral_embedding_unknown_affinity(seed=36):
-    """Test that SpectralClustering fails with an unknown affinity type"""
-    centers = np.array([
-        [0., 0., 0.],
-        [10., 10., 10.],
-        [20., 20., 20.],
-    ])
-    X, true_labels = make_blobs(n_samples=100, centers=centers,
-                                cluster_std=1., random_state=42)
-    D = rbf_kernel(X)  # Distance matrix
-
-    se_precomp = SpectralEmbedding(n_components=1, affinity="<unknown>",
-                                   random_state=np.random.RandomState(seed))
-    assert_raises(ValueError, se_precomp.fit, S)
+                spectral_clustering.steps[1][1].labels_,
+                SC.labels_), 0.0, 0)
